Delete PVC alongside helm uninstall
Helm does not currently delete PVCs meaning
they build up uneccesarily, eating up cluster
resource and cost.

This change implements a delete of any
PVC containing the release name.

It adds input flag/option that needs to be supplied
with value true. This means the version can be released
without changing existing behaviour as the deletion
of PVCs is very destructive and we should probably
not assume all use cases need it, even for UAT
environments. A major version bump might be an alternative
way to handle this.